Floating mechanism implementation
Combine the ship control and the stylized water assets package with the Align Transform To Water function. Because the "Align Transform To Water" script, in order to make the boat fit the water waves, will calculate a new rotation angle in real time. So in order to achieve the desired result. A new small script has been added to force the ship model to maintain its control orientation while remaining in close contact with the water surface. The new "VisualsYawLock" is responsible for making the ship follow the turns.